requirements.txt
openai
openai-agents
youtube-transcript-api
python-dotenv